# Refactor reclist to use deque rather than vector