--- a/inst/include/datarecord.h
+++ b/inst/include/datarecord.h
@@ -87,11 +87,11 @@ public:
   
   void ii(double ii_){Ii = ii_;}
   double ii(){return Ii;}
-
+  
   double fn(){return Fn;}
   void fn(double fn_){Fn = fn_;}
-
-  void schedule(std::deque<rec_ptr>& thisi, double maxtime, bool put_ev_first, 
+  
+  void schedule(reclist& thisi, double maxtime, bool put_ev_first, 
                 const unsigned int maxpos, double lagt);
   void implement(odeproblem* prob);
   void steady_zero(odeproblem* prob, LSODA& solver);
@@ -118,13 +118,13 @@ public:
   bool is_phantom() {return !Output && !Fromdata;}
   bool is_lagged() {return Lagged;}
   void lagged() {Lagged = true;}
-
+  
   int Pos; ///< record position number
   unsigned short int Evid; ///< record event ID
   unsigned short int Ss; ///< record steady-state indicator
   short int Cmt; ///< record compartment number
   unsigned int Addl; ///< number of additional doses
-
+  
   double Time; ///< record time
   double Id; ///< record ID value
   double Amt; ///< record dosing amount value
@@ -161,7 +161,7 @@ struct CompRec {
   }
 };
 
-void insert_record(std::deque<rec_ptr>& thisi, const int start, rec_ptr& rec, 
+void insert_record(reclist& thisi, const int start, rec_ptr& rec, 
                    const bool put_ev_first);
 
 #endif