--- a/inst/include/odeproblem.h
+++ b/inst/include/odeproblem.h
@@ -24,6 +24,7 @@
 #define ODEPROBLEM_H
 #include <math.h>
 #include <vector>
+#include <deque>
 #include "RcppInclude.h"
 #include "mrgsolv.h"
 #include "datarecord.h"
@@ -38,7 +39,7 @@
 class odeproblem;
 
 //! vector of <code>datarecord</code> objects for one <code>ID</code>
-typedef std::vector<rec_ptr> reclist;
+typedef std::deque<rec_ptr> reclist;
 
 //! vector of <code>reclist</code> vectors comprising  a data set
 typedef std::vector<reclist> recstack;