--- a/src/devtran.cpp
+++ b/src/devtran.cpp
@@ -185,7 +185,7 @@ Rcpp::List DEVTRAN(const Rcpp::List parin,
   // Find tofd
   std::vector<double> tofd;
   if(tad) {
-    //tofd.reserve(a.size());
+    tofd.reserve(a.size());
     for(recstack::const_iterator it = a.begin(); it !=a.end(); ++it) {
       for(reclist::const_iterator itt = it->begin(); itt != it->end(); ++itt) {
         if((*itt)->is_dose()) {
@@ -250,7 +250,7 @@ Rcpp::List DEVTRAN(const Rcpp::List parin,
       
       reclist z;
       
-      //z.reserve(tgridn[i]);
+      //z.reserve(tgridn[i]); // TODO: remove
       
       for(int j = 0; j < tgridn[i]; ++j) {
         rec_ptr obs = NEWREC(tgrid(j,i),nextpos,true);
@@ -272,7 +272,7 @@ Rcpp::List DEVTRAN(const Rcpp::List parin,
         n = tgridn.at(0);
       } 
       
-      //it->reserve((it->size() + n));
+      //it->reserve((it->size() + n)); // TODO: remove
       for(int h=0; h < n; ++h) {
         it->push_back(designs[tgridi[j]][h]);
         ++obscount;