--- a/src/mrgsolve.cpp
+++ b/src/mrgsolve.cpp
@@ -382,7 +382,7 @@ Rcpp::List EXPAND_OBSERVATIONS(
   obscount = 0;
   
   int n_time = int(times.size());
-  std::deque<rec_ptr> z(n_time);
+  reclist z(n_time);
   for(int j = 0; j < n_time; ++j) {
     rec_ptr obs = NEWREC(times[j],nextpos,true);
     z[j] = obs;