@kyleam 2024-04-19
> Do you know of any way to profile this branch (`refactor/deque`) against, say, `refactor/alag-insert` to understand the impact?

I think we can use Valgrind's [Massif](https://valgrind.org/docs/manual/ms-manual.html) heap profiler.

Here's your script from above, removing the timing statements and output comments.

<details>
<summary>script</summary>

```R
code <- '
$PARAM @annotated
CL   :  1 : Clearance (volume/time)
V2   : 20 : Central volume (volume)
Q    :  2 : Inter-compartmental clearance (volume/time)
V3   : 10 : Peripheral volume of distribution (volume)
KA   :  1 : Absorption rate constant (1/time)
LAG  :  0 : Absorption lag time

$CMT @annotated
EV     : Extravascular compartment (mass)
CENT   : Central compartment (mass)
PERIPH : Peripheral compartment (mass)

$GLOBAL
#define CP (CENT/V2)

$MAIN
ALAG_EV = LAG;

$PKMODEL ncmt = 2, depot = TRUE

$CAPTURE @annotated
CP : Plasma concentration (mass/time)
'

library(mrgsolve)
library(dplyr)

mod <- mcode("model", code)

mmod <- param(mod, LAG = 0.1)

e <- ev(amt = 100, ii = 24, addl = 9)
mrgsim(mod,  e, end = 240)
mrgsim(mmod, e, end = 240)

inf <- mutate(e, rate = 990)
mrgsim(mod,  inf, end = 240)
mrgsim(mmod, inf, end = 240)


e2 <- ev_rep(e, 1:1000) %>% mutate(addl = 27)
  mrgsim(mod, e2, end = 24*28)
  mrgsim(mmod, e2, end = 24*28)

inf2 <- ev_rep(inf, 1:1000) %>% mutate(addl = 27)
mrgsim(mod, inf2, end = 24*28)
mrgsim(mmod, inf2, end = 24*28)

dd <- expand.ev(amt = 100, ii = 24, addl = 300, ID = 1:100)
ddd <- lapply(1:100, function(i) dd) %>% bind_rows()

mrgsim(mod, ddd, end = 5000, delta = 24)

mrgsim(mmod, ddd, end = 5000, delta = 24)

ddd <- realize_addl(ddd)

mrgsim(mod, ddd, end = 5000, delta = 24)

mrgsim(mmod, ddd, end = 5000, delta = 24)
```

</details>

And then on the tip of this PR (864a432f), I did

```
$ make
$ R -d "valgrind --tool=massif --massif-out-file=massif.out.pr" -s -e 'source("do.R")'
```

(If on Metworx, you'll need to install valgrind: `sudo apt-get install valgrind`.)

And similar for the tip of `refactor/alag-insert` (573e9c70):

```
# check out refactor/alag-insert
$ make
$ R -d "valgrind --tool=massif --massif-out-file=massif.out.base" -s -e 'source("do.R")'
```

Here are some exported `massif-visualizer` plots (need to install on Metworx).

**refactor/alag-insert**

![massif-base](https://github.com/metrumresearchgroup/mrgsolve/assets/1297788/b8f0ddd2-79fb-4cb0-9b9d-f6c14d57b1b4)

**this PR**

![massif-pr](https://github.com/metrumresearchgroup/mrgsolve/assets/1297788/9d285e60-9899-408d-9c3a-2891d659e7d2)



---

Anyway, I'm not sure how useful those particular images are (also visualizer UI is more helpful than static image) or whether that script is a good thing to profile, but let me know if you think that's looking like the sort of info you want and we can explore more.