@kylebaron 2024-04-19
Thanks a lot, @kyleam . This is super cool. I was able to run with valgrind on Metworx with a reduced script (just the very last problem ... with larger data set). I installed the visualizer, but wasn't sure how / if I could invoke this from the comand line. 

Anyway, this seems to say that the allocations are pretty close between the vector and the deque?  Maybe a small difference but it didn't increase by 50 or 100% or anything like that. Is that the way you're viewing this?