@kyleam 2024-04-19
> I installed the visualizer, but wasn't sure how / if I could invoke this from the comand line.

Ah, sorry, I meant to mention that.  I'm sure you could set up some port forwarding or whatnot to see it via laptop, but I just invoked it from the command line (make sure `DISPLAY=:99` is set in environment and then went to the Metworx desktop).