@kyleam 2024-04-19
> Maybe a small difference but it didn't increase by 50 or 100% or anything like that. Is that the way you're viewing this?

Yes, that was my initial take.